+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 _This is the changelog for the entire project_
 
+## 8/31/2024
+
+- Updated Button in GymUI to have sizes
+
 ## 8/23/2024
 
 - Updated CI test to not run on draft pull requests